refactor: criado um componente para renderizar os botões de filtro para as telas de done e favorites recipes
